Move (base-)views to the browser package.
