+++ opengever/tabbedview/__init__.py
@@ -1,11 +1,11 @@
-from opengever.tabbedview.base_tabs import BaseCatalogListingTab  # noqa
-from opengever.tabbedview.base_tabs import BaseListingTab  # noqa
-from opengever.tabbedview.base_tabs import GeverTabMixin  # noqa
 from opengever.tabbedview.basesource import GeverTableSource  # noqa
+from opengever.tabbedview.browser.base_tabs import BaseCatalogListingTab  # noqa
+from opengever.tabbedview.browser.base_tabs import BaseListingTab  # noqa
+from opengever.tabbedview.browser.base_tabs import GeverTabMixin  # noqa
+from opengever.tabbedview.browser.tabbed import GeverTabbedView  # noqa
+from opengever.tabbedview.browser.tabbed import ModelProxyTabbedView  # noqa
 from opengever.tabbedview.catalog_source import GeverCatalogTableSource  # noqa
 from opengever.tabbedview.sqlsource import SqlTableSource  # noqa
-from opengever.tabbedview.tabbed import GeverTabbedView  # noqa
-from opengever.tabbedview.tabbed import ModelProxyTabbedView  # noqa
 from zope.i18nmessageid import MessageFactory
 import logging
